+++ CHANGELOG.md
@@ -8,6 +8,7 @@ project groups.
 - Improved parsing for IAR EW toolchain messages as diagnostic problems.
 - Improved parsing for SDCC toolchain messages as diagnostic problems.
 - Implemented parsing for Watcom toolchain messages as diagnostic problems.
+- Implemented parsing for Clang toolchain messages as diagnostic problems.
 
 ## 2.1.1